requirements.txt
aerich==0.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiomysql==0.2.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncio==3.4.3
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dashscope==1.24.0
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pip==25.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
tortoise-orm==0.25.1
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
wheel==0.45.1
yarl==1.20.1
